protected MVC architecture by moving routes to controllers and cleaned up an import
