+++ controllers/index.js
@@ -1,7 +1,9 @@
 const router = require('express').Router();
 
-const apiRoutes = require('./api/home');
+const apiRoutes = require('./api');
+const homeRoutes = require('./home');
 
 router.use('/api', apiRoutes);
+router.use('/', homeRoutes);
 
 module.exports = router;
